Updated error handling in UInput
if UInput.capabilities() fails it is due to a missing UInput.device
this can have two causes:
1. the system was to slow to create the /dev/input/event* file
2. wrong permissions on the event* file

The first cause is handled by trying to find the event* file, but only once.
We probably can not hanlde the second cause, so we assume the event can be handled
this might result in missing events if our assumption was wrong.